requirements.txt
pandas
tabula-py
openpyxl
pyinstaller